refactor(mcp): tighten module visibility and co-locate tests
Three cleanup items from the /age review of the module split:

1. Tool entry points (tool_definitions, tool_deps, tool_diff, tool_edit,
   tool_files, tool_read, tool_search, tool_session) plus parse_file_edit
   narrowed from pub(crate) to pub(in crate::mcp). Their pub(super) use
   re-exports in src/mcp/tools/mod.rs likewise narrowed. The originals in
   the monolithic src/mcp.rs were truly private; pub(in crate::mcp) is the
   tightest visibility that keeps src/mcp/mod.rs's `use tools::tool_*`
   imports working.

2. Services struct, its impl methods, and dispatch_tool in src/mcp/mod.rs
   dropped pub(crate) — only callers live in mod.rs itself, so default
   (module-private) visibility is the right scope.

3. Tests co-located with the code they exercise:
   - resolve_scope_* and scope_flag_overrides_bad_cwd → src/mcp/tools/mod.rs
   - tool_files_* (+ scratch_project helper) → src/mcp/tools/files.rs
   - parse_file_edit_rejects_empty_edits_array → src/mcp/tools/edit.rs
   The extract_root_* tests and package_root_finds_project_from_subdirectory
   stay in src/mcp/mod.rs since they exercise mod.rs functions. The
   #[cfg(test)] use tools::{parse_file_edit, resolve_scope} import in
   mod.rs and the matching pub(super) use re-export in tools/mod.rs are
   removed since neither name is referenced from mod.rs anymore.

No behavior changes; 366 tests pass.
