+++ src/mcp/tools/mod.rs
@@ -7,23 +7,21 @@ mod read;
 mod search;
 mod session;
 
-pub(crate) use definitions::tool_definitions;
-pub(crate) use deps::tool_deps;
-pub(crate) use diff::tool_diff;
-#[cfg(test)]
-pub(crate) use edit::parse_file_edit;
-pub(crate) use edit::tool_edit;
-pub(crate) use files::tool_files;
-pub(crate) use read::tool_read;
-pub(crate) use search::tool_search;
-pub(crate) use session::tool_session;
+pub(super) use definitions::tool_definitions;
+pub(super) use deps::tool_deps;
+pub(super) use diff::tool_diff;
+pub(super) use edit::tool_edit;
+pub(super) use files::tool_files;
+pub(super) use read::tool_read;
+pub(super) use search::tool_search;
+pub(super) use session::tool_session;
 
 use std::path::PathBuf;
 
 use serde_json::Value;
 
 /// Falls back to cwd when scope is invalid, with a warning message.
-pub(crate) fn resolve_scope(args: &Value) -> (PathBuf, Option<String>) {
+pub(super) fn resolve_scope(args: &Value) -> (PathBuf, Option<String>) {
     let raw_str = args.get("scope").and_then(|v| v.as_str()).unwrap_or(".");
     let raw: PathBuf = raw_str.into();
     let resolved = raw.canonicalize().unwrap_or_else(|_| raw.clone());
@@ -46,5 +44,88 @@ pub(super) fn apply_budget(output: String, budget: Option<u64>) -> String {
     match budget {
         Some(b) => crate::budget::apply(&output, b),
         None => output,
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn resolve_scope_explicit_arg() {
+        let tmp = tempfile::tempdir().unwrap();
+        let args = serde_json::json!({ "scope": tmp.path().to_str().unwrap() });
+        let (scope, warning) = resolve_scope(&args);
+        assert_eq!(scope, tmp.path().canonicalize().unwrap());
+        assert!(warning.is_none());
+    }
+
+    #[test]
+    fn resolve_scope_no_arg_uses_cwd() {
+        let args = serde_json::json!({});
+        let (scope, warning) = resolve_scope(&args);
+        // With no arg, defaults to "." which is cwd
+        let cwd = std::env::current_dir().unwrap();
+        // The function returns "." when resolved == cwd
+        assert!(scope == PathBuf::from(".") || scope == cwd);
+        assert!(warning.is_none());
+    }
+
+    #[test]
+    fn resolve_scope_invalid_dir_warns() {
+        let args = serde_json::json!({ "scope": "/nonexistent/directory/zzz" });
+        let (scope, warning) = resolve_scope(&args);
+        assert_eq!(scope, PathBuf::from("."));
+        assert!(warning.is_some());
+        assert!(warning.unwrap().contains("not a valid directory"));
+    }
+
+    /// Reproduces issue #37: MCP host launches tilth with cwd=/. The --scope
+    /// flag should override this.
+    #[test]
+    fn scope_flag_overrides_bad_cwd() {
+        let project = tempfile::tempdir().unwrap();
+        let project_path = project.path();
+
+        // Create a manifest so package_root can find it
+        std::fs::write(
+            project_path.join("Cargo.toml"),
+            "[package]\nname = \"test\"",
+        )
+        .unwrap();
+        std::fs::create_dir(project_path.join("src")).unwrap();
+        std::fs::write(project_path.join("src/main.rs"), "fn main() {}").unwrap();
+
+        // Save current cwd
+        let orig_cwd = std::env::current_dir().unwrap();
+
+        // Simulate Codex: cwd=/
+        std::env::set_current_dir("/").unwrap();
+
+        // Without --scope: resolve_scope returns "." which is /
+        let args = serde_json::json!({});
+        let (scope, _) = resolve_scope(&args);
+        assert_eq!(
+            scope,
+            PathBuf::from("."),
+            "Without --scope, should return . (which is /)"
+        );
+
+        // With --scope pointing to project: set_current_dir should fix everything
+        let _ = std::env::set_current_dir(project_path);
+        let args = serde_json::json!({});
+        let (scope, _) = resolve_scope(&args);
+        assert_eq!(
+            scope,
+            PathBuf::from("."),
+            "After chdir to project, . should resolve correctly"
+        );
+
+        // Verify tilth_files would search in the project, not /
+        let cwd = std::env::current_dir().unwrap();
+        assert_eq!(cwd, project_path.canonicalize().unwrap());
+
+        // Restore
+        std::env::set_current_dir(orig_cwd).unwrap();
     }
 }
